Refactor tuples and anonymous structs to be unified with regular structs
Previously, (some) tuples and anonymous structs were represented using
different `Type` tags to regular structs. Because these were implemented
separately, it led to many bugs around these constructs:
* explicit tuple types (`struct { u32 }`) were completely distinct from implicit tuple types (from e.g. concatenation)
* optimized struct layout did not work for anonymous structs and (most) tuples
* tuples lost their layout when performing operations like concatenation
* some coercions were broken and/or gave nonsensical errors
* tuple type equality checks did not work how they were meant to
* tuple concatenation and other operations did not properly check for comptime-known values
* some more stuff I've forgotten by now

This commit also incidentally fixes some other issues:
* `default_value` wasn't resolved correctly in struct reification
* coercions between anon struct types were broken
* error printing of tuple and anon struct types used non-canonical syntax
* some more stuff I've forgotten by now

This change is a net decrease in compiler LOC: a lot of Sema logic was
able to be simplified, and backends no longer have to worry about
special handling for `tuple` and `anon_struct` tags. This in particular
can be a big win for the C backend, whose lowering of types can likely
be simplified a lot. This PR does not make this change optimally: I'm
leaving that work to @jacobly0 since he's more familiar with CBE's type
lowering logic.

Here's an overview of the semantics implemented by this change. We may
end up wishing to change these in some ways - tuples were not
well-specified so a lot of choices here were made based on my own
opinion of what makes sense - but they're certainly an improvement on
the status quo.

Tuples are similar to structs, but have some extra limitations:
* tuples may not contain decls, comptime blocks, or tests
* non-`comptime` tuple fields may not have default values

The empty struct `struct {}` is a tuple type, but if decls are added to
it, it is a normal struct (namespace) type.

Tuple types use structural equality. This means two tuple types `S` and
`T` are equivalent if they have the same layout and the same fields (and
all comptime fields have the same values). Tuples are not lazy like
regular structs: for instance, `const S = struct { *S };` is not a valid
declaration, and will trigger a dependency loop error. This is necessary
for type deduplication to work: in this sense, tuples are more like
arrays than structs.

Note that the use of structural equality means that tuples are not
attached to any particular source location. Error messages referencing
them print the type in full, e.g. `extern struct { comptime u32 = 3 }`.

For any struct or tuple type, there is an equivalent "anonymous" type,
which is the type of a literal like `.{3}` or `.{ .x = 5 }` with no
know result type. Anonymous types can coerce to any structurally
equivalent type where the field types are all coercible; for instance,
`.{3}` can coerce to a `struct { comptime comptime_int = 3 }`, or to a
`struct { u32 }`, an `extern struct { u32 }`, etc. `.{}` is of an
anonymous tuple type, but as a special case can also coerce to standard
structs. Like tuples, anonymous struct types are printed in full (since
their names are automatically generated and unhelpful), but they also
have a source location attached, which is the location of the anonymous
literal.

Tuples have two special operations: concatenation (`++`) and
multiplication (`**`). The former operation requires both structs to
have the same layout (but not the same backing int type if packed). If
the layout is packed, the result's backing int type is implicit. The
result is never an anonymous type.

The current implementation has some issues regarding incremental
compilation. As far as I'm aware, they shouldn't *break*, but it is
possible for unnecessary recompilation to be performed. Currently, the
anonymous decls which are created to contain tuple types are parented to
the decl which first used the tuple. This means the deletion of this
somewhat arbitrary parent declaration may trigger an arbitrary amount of
unnecessary re-analysis. This isn't ideal, but tuple types are used
rarely enough that it shouldn't be a major issue for now. Nevertheless,
this should definitely be improved upon in a follow-up PR.

Resolves: #15393
Resolves: #14612