Check the file type after open() in os.openFile() when using WASI
When the `O_DIRECTORY` flag is not present, Wasmtime now opens
files and directories the same way, and returns a descriptor in
any case.

So, explicitly check the file type after opening it, and return
`err.IsDir` if what we actually have is a directory.

Fixes the `lib/std/fs` `"file operations on directories"` test
on WASI with Wasmtime 9.0.0