Fix std.hash benchmarks
The benchmarks were reusing the same input buffer over and over again,
giving a chance to the optimizer for computing the hash only once,
and multiplying the output by the number of iterations.

After this, the results are *very* different (not in a good way, but
at least, they are closer to reality).

Also add xxhash and siphash.

xxhash hash() function had a different API than all other std.hash
function that forced the key to be 0. Make it consistent with other
functions, especially since this is not a cryptographic hash function.

Remove some useless `pub` and declarations from siphash by the way.