UEFI main: remove support for usize return type
I know this is kind of abrupt and there's no deprecation either but I
don't think many people used this anyway.
The reason I removed it is because std.os.uefi.Status is a usize too
and is a non-exhaustive enum which means you can @intToEnum any value to
std.os.uefi.Status if you want to return something outside of the
standard status values.

I suppose returning non-specified status values is discouraged so this
would add to that.

This change will make more sense in the following commits.
The purpose is mainly unification.