std: make exit API cross-platform
The motivation for this change is that `std.os.exit(int)` **is not cross-platform**.
The usual assumption is that the OS allows returning an integer upon
exiting which actually is in the range 0 to 255. This is not the case on
all operating systems. We cannot assume u8 as the exit status type on all systems.

Therefore we make the API cross-platform with this change.
One example is Plan 9. On Plan 9 you can't return exit codes like on
Linux. Instead, you return strings. This is now supported.
See the TODO for this that is now resolved.

I have confirmed this to work on the UEFI with this code:
```
pub fn main() !@import("std").os.uefi.Status {
    @import("std").process.exit(.Aborted);
    return error.SomethingWentWrong;
}
```

I've also tested this on Linux and Windows.

I was not able to test this on Plan 9 unfortunately because I don't
understand how to pass a compiled binary through QEMU to run it on Plan 9.
I'm quite confident any regressions will be easy to fix though.

Indeed, this means that if you write cross-platform code, you should now
use `std.os.exit(std.os.exit_status_success)` instead of `std.os.exit(0)`.
And similarly `std.os.exit(exit_status_failure)` instead of `std.os.exit(1)`.
I made some but not all invocations of `std.os.exit` to be
cross-platform but it's not too important right now because currently
no platform that doesn't use integers as the exit status runs on the CI
or is one of the commonly used platforms. Replacing the rest can be done
with time.

---

In start.zig, we now also use the same "bad main return type" error message everywhere.
This is possible thanks to a previous commit that removes support for usize as the return type for main on the UEFI.

Regression: one thing that's slightly worse after this is the bad_main_ret error message on the UEFI.
For UEFI programs it now prints "expected return type of main to be 'void', '!void, 'os.uefi.status.Status', '!os.uefi.status.Status', or 'noreturn'"
Previously it was "std.os.uefi.Status". Now it's "os.uefi.status.Status". This is a problem caused by using @typeName.
One way to solve this would be to add a function to std.meta or similar that creates a type string using its own logic.
It could skip private namespaces (like os.uefi.status).
This is a rather minor problem though and I think it's okay for now.

On other platforms like Linux the error message looks pretty good:
```
$ cat x.zig
const std = @import("std");

pub fn main() !u123 {
    return 126;
}
$ zig run x.zig --zig-lib-dir lib
lib/std/start.zig:582:29: error: expected return type of main to be 'void', '!void, 'u8', '!u8', or 'noreturn'
                    else => @compileError(bad_main_ret),
                            ^~~~~~~~~~~~~~~~~~~~~~~~~~~
```
As for `if (std.os.ExitStatus == void)` in the bad_main_ret definition,
std.os.ExitStatus can currently not be `void` but in the future that will almsot certainly change.
I'm very confident there will be an operating system that you cannot pass a value to when exiting.