Add `os.windows.nls.upcaseW`, a cross-platform implementation of `RtlUpcaseUnicodeChar`
This allows doing Windows-style case insensitive comparisons from any target, but means that it will need to include its own copy of the uppercase data table (5,088 bytes) to do so.

When targeting Windows, the ntdll functions are used instead to avoid including a redundant copy of the uppercase data in the resulting binary.